Refactor packages o.b.lime.set_statistics to o.b.l.statistics and o.b.lime.set_theory to o.b.l.op.
