+++ lime-cli/src/main/scala/org/bdgenomics/lime/cli/Intersection.scala
@@ -20,7 +20,7 @@ package org.bdgenomics.lime.cli
 import org.apache.spark.SparkContext
 import org.bdgenomics.adam.rdd.ADAMContext._
 import org.bdgenomics.adam.rdd.ADAMSaveAnyArgs
-import org.bdgenomics.lime.set_theory.ShuffleIntersection
+import org.bdgenomics.lime.op.ShuffleIntersection
 import org.bdgenomics.utils.cli._
 import org.kohsuke.args4j.Argument
