+++ lime-core/src/main/scala/org/bdgenomics/lime/op/Cluster.scala
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.bdgenomics.lime.set_theory
+package org.bdgenomics.lime.op
 
 import org.bdgenomics.adam.models.ReferenceRegion
 import org.bdgenomics.adam.rdd.{ GenomicRDD, GenericGenomicRDD }
