+++ lime-core/src/test/scala/org/bdgenomics/lime/op/MergeSuite.scala
@@ -15,18 +15,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.bdgenomics.lime.set_statistics
+package org.bdgenomics.lime.op
 
 import org.bdgenomics.adam.rdd.ADAMContext._
 import org.bdgenomics.lime.LimeFunSuite
 
-class JaccardSuite extends LimeFunSuite {
-  sparkTest("test jaccard distance between two regions") {
-    val leftGenomicRDD = sc.loadFeatures(resourcesFile("/intersect_with_overlap_00.bed")).sortLexicographically()
-    val rightGenomicRdd = sc.loadFeatures(resourcesFile("/intersect_with_overlap_01.bed")).sortLexicographically()
-
-    val jaccard_dist = new JaccardDistance(leftGenomicRDD, rightGenomicRdd).compute()
-
-    assert(JaccardStatistic(9240, 32917, 0.28070601816690466, 3).equals(jaccard_dist))
+class MergeSuite extends LimeFunSuite {
+  sparkTest("test local merge when all data merges to a single region") {
+    val genomicRdd = sc.loadBed(resourcesFile("/cpg_20merge.bed"))
+    val x = ShuffleMerge(
+      genomicRdd)
+      .compute()
+    assert(x.rdd.count == 1)
   }
 }
